fix(z2m): accept fractional voltage/illuminance/voc/transmit_power
Smart plugs with power metering (e.g. Third Reality 3RSP02028BZ) report
sub-integer voltage/current/power/frequency. The ZigbeeMessage struct had
Voltage as *int, so json.Unmarshal aborted on "voltage":120.8 and the
controller could neither register the device nor route its messages.

Switch Voltage, Illuminance, VOC, and TransmitPower to *float64 to match
z2m's schema (numeric, not integer). All four are already cast to float64
at the metric setter (Set(float64(*m.X))), so this is purely a decode-side
relaxation — no downstream changes needed.

Add TestZigbeeMessageFractionalNumerics with a real 3RSP02028BZ payload
and TestZigbeeMessageFloatVOCAndIlluminance for the same risk class on
VOC and Illuminance.

+++ pkg/iot/routers/zigbee2mqtt/message.go
@@ -22,16 +22,16 @@ type ZigbeeMessage struct {
 	Co2           *float64 `json:"co2,omitempty"`
 	Formaldehyde  *float64 `json:"formaldehyd,omitempty"`
 	Humidity      *float64 `json:"humidity,omitempty"`
-	Illuminance   *int     `json:"illuminance,omitempty"`
+	Illuminance   *float64 `json:"illuminance,omitempty"`
 	LinkQuality   *int     `json:"linkquality,omitempty"`
 	Occupancy     *bool    `json:"occupancy,omitempty"`
 	SoilMoisture  *float64 `json:"soil_moisture,omitempty"`
 	State         *string  `json:"state,omitempty"`
 	Tamper        *bool    `json:"tamper,omitempty"`
 	Temperature   *float64 `json:"temperature,omitempty"`
-	TransmitPower *int     `json:"transmit_power,omitempty"`
-	VOC           *int     `json:"voc,omitempty"`
-	Voltage       *int     `json:"voltage,omitempty"`
+	TransmitPower *float64 `json:"transmit_power,omitempty"`
+	VOC           *float64 `json:"voc,omitempty"`
+	Voltage       *float64 `json:"voltage,omitempty"`
 	WaterLeak     *bool    `json:"water_leak,omitempty"`
 }
